requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.10.0
torch==2.1.2
torchvision==0.16.2
transformers==4.37.2
pillow==10.2.0
opencv-python==4.9.0.80
librosa==0.10.1
moviepy==1.0.3
python-dotenv==1.0.0